Move graph-layer out of src/layers/core, and repackage everything within examples/graph/graph-layer.
