+++ examples/graph/webpack.config.js
@@ -16,7 +16,7 @@ module.exports = {
       {
         // Compile ES2015 using buble
         // test: /\.js$/,
-        test: /^((?!graph-layout-layer).)*\.js$/,
+        test: /^((?!graph-layer).)*\.js$/,
         loader: 'buble-loader',
         include: [resolve('.')],
         exclude: [/node_modules/],
@@ -25,12 +25,12 @@ module.exports = {
         }
       },
       {
-        // ...except graph-layout-layer.js, which needs to not be transpiled
+        // ...except graph-layer.js, which needs to not be transpiled
         // when running this example within deck.gl source,
-        // so that it can extend the untranspiled GraphLayer from deck.gl src.
+        // so that it can extend the untranspiled CompositeLayer from deck.gl src.
         // TODO: when running against npm-packaged deck.gl (outside of deck.gl src),
-        // graph-layout-layer.js should be transpiled like everything else...
-        test: /graph-layout-layer/,
+        // graph-layer.js should be transpiled like everything else...
+        test: /graph-layer/,
         loader: 'exports-loader'
       }
     ]
